moving redux logic out of gamelist component into container. also moving actions and reducer into the container folder
