+++ src/configureStore.js
@@ -1,7 +1,7 @@
 import thunkMiddleware from 'redux-thunk';
 import createLogger from 'redux-logger';
 import { createStore, applyMiddleware } from 'redux';
-import rootReducer from './reducers';
+import rootReducer from './rootReducer';
 
 const loggerMiddleware = createLogger();
