+++ src/containers/gameList/constants.js
@@ -1,0 +1,1 @@
+export const RECEIVE_GAMELIST = 'RECEIVE_GAMELIST';
