+++ src/containers/gameList/reducer.js
